# api/swagger/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "description": "This is a client API for Wice CRM",
    "version": "0.0.1",
    "title": "Wice CRM API",
    "termsOfService": "https://wice.de",
    "contact": {
      "email": "syanev@wice.de"
    },
    "license": {
      "name": "",
      "url": ""
    }
  },
...
  "security": [{
    "ApiKeyAuth": [],
    "CookieAuth": []
  }],
  "paths": {
...
    "/api/v1/persons": {
      "get": {
        "tags": [
          "person"
        ],
        "x-swagger-router-controller": "person",
        "summary": "Get all persons",
        "description": "Returns all available persons",
        "responses": {
          "200": {
            "description": "An array of all persons objects",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/getPersonResponse"
                }
              }
            }
          },
          "401": {
            "description": "Invalid input"
          }
        }
      },
      "post": {
        "tags": [
          "person"
        ],
        "x-swagger-router-controller": "person",
        "summary": "Create a person",
        "description": "Returns the new created person.",
        "responses": {
          "200": {
            "description": "Supplies the new created person",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/createPersonResponse"
                }
              }
            }
          },
          "401": {
            "description": "Invalid input"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/createPersonRequest"
              }
            }
          },
          "description": "First name and last name which will be passed on to the server",
          "required": true
        }
      }
    },
    "/api/v1/persons/{rowid}": {
      "get": {
        "tags": [
          "person"
        ],
        "x-swagger-router-controller": "person",
        "summary": "Get a person by a given rowid",
        "description": "Returns the single person ",
        "parameters": [{
          "name": "rowid",
          "in": "path",
          "description": "Person's rowid",
          "required": true,
          "schema": {
            "type": "string",
            "format": "string"
          }
        }],
        "responses": {
          "200": {
            "description": "Person object",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/getPersonResponse"
                }
              }
            }
          },
          "401": {
            "description": "Invalid input"
          }
        }
      },
      "put": {
        "tags": [
          "person"
        ],
        "x-swagger-router-controller": "person",
        "summary": "Update a person",
        "description": "Returns updated person",
        "parameters": [{
          "name": "rowid",
          "in": "path",
          "description": "The rowid of the person to be deleted",
          "required": true,
          "schema": {
            "type": "string",
            "format": "string"
          }
        }],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": ""
              }
            }
          },
          "description": "Some description.",
          "required": true
        },
        "responses": {
          "200": {
            "description": "An array of all persons objects",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": ""
                }
              }
            }
          },
          "404": {
            "description": "An array with false for each not found uid",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "example": ""
                }
              }
            }
          },
          "401": {
            "description": "Invalid input"
          }
        }
      },
      "delete": {
        "tags": [
          "person"
        ],
        "x-swagger-router-controller": "person",
        "summary": "Delete a person",
        "description": "Returns deleted person's rowid",
        "parameters": [{
          "name": "rowid",
          "in": "path",
          "description": "The id of the user to be deleted",
          "required": true,
          "schema": {
            "type": "string",
            "format": "string"
          }
        }],
        "responses": {
          "200": {
            "description": "An array of all persons objects",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": ""
                }
              }
            }
          },
          "404": {
            "description": "An array with false for each not found uid",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "example": ""
                }
              }
            }
          },
          "401": {
            "description": "Invalid input"
          }
        }
      }
    },
    "/api/v1/organizations": {
      "get": {
        "tags": [
          "organization"
        ],
        "x-swagger-router-controller": "organization",
        "summary": "Get all organizations",
        "description": "Returns all available organizations",
        "responses": {
          "200": {
            "description": "An array of all organizations objects",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": ""
                }
              }
            }
          },
          "401": {
            "description": "Invalid input"
          }
        }
      },
      "post": {
        "tags": [
          "organization"
        ],
        "x-swagger-router-controller": "organization",
        "summary": "Create an organization",
        "description": "Returns the new created organization.",
        "responses": {
          "200": {
            "description": "Supplies the new created organization",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": ""
                }
              }
            }
          },
          "401": {
            "description": "Invalid input"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/createOrganizationRequest"
              }
            }
          },
          "description": "Organization's name which will be passed on to the server",
          "required": true
        }
      }
    },
    "/api/v1/organizations/{rowid}": {
      "get": {
        "tags": [
          "organization"
        ],
        "x-swagger-router-controller": "organization",
        "summary": "Get an organization by a given rowid",
        "description": "Returns the single organization ",
        "parameters": [{
          "name": "rowid",
          "in": "path",
          "description": "Organization's rowid",
          "required": true,
          "schema": {
            "type": "string",
            "format": "string"
          }
        }],
        "responses": {
          "200": {
            "description": "Organization object",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": ""
                }
              }
            }
          },
          "401": {
            "description": "Invalid input"
          }
        }
      },
      "put": {
        "tags": [
          "organization"
        ],
        "x-swagger-router-controller": "organization",
        "summary": "Update an organization",
        "description": "Returns updated organization",
        "parameters": [{
          "name": "rowid",
          "in": "path",
          "description": "The rowid of the organization to be deleted",
          "required": true,
          "schema": {
            "type": "string",
            "format": "string"
          }
        }],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": ""
              }
            }
          },
          "description": "Some description.",
          "required": true
        },
        "responses": {
          "200": {
            "description": "An array of all organization objects",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": ""
                }
              }
            }
          },
          "404": {
            "description": "An array with false for each not found uid",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "example": ""
                }
              }
            }
          },
          "401": {
            "description": "Invalid input"
          }
        }
      },
      "delete": {
        "tags": [
          "organization"
        ],
        "x-swagger-router-controller": "organization",
        "summary": "Delete an organization",
        "description": "Returns deleted organization's rowid",
        "parameters": [{
          "name": "rowid",
          "in": "path",
          "description": "The id of the organization to be deleted",
          "required": true,
          "schema": {
            "type": "string",
            "format": "string"
          }
        }],
        "responses": {
          "200": {
            "description": "An array of all organization objects",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": ""
                }
              }
            }
          },
          "404": {
            "description": "An array with false for each not found uid",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "example": ""
                }
              }
            }
          },
          "401": {
            "description": "Invalid input"
          }
        }
      }
    },
...
  },
...
  "servers": [{
    "url": "http://localhost:5000/"
  }],
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-KEY"
      },
      "CookieAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "wice-cookie"
      }
    },
    "schemas": {
...
      "getPersonResponse": {
        "type": "array",
        "items": {
          "type": "object",
          "properties": {
            "rowid": {
              "type": "string",
              "example": "821528"
            },
            "mandant": {
              "type": "string",
              "example": "821528"
            },
            "for_rowid": {
              "type": "string",
              "example": "821528"
            },
            "same_contactperson": {
              "type": "string",
              "example": "821528"
            },
            "name": {
              "type": "string",
              "example": "821528"
            },
            "firstName": {
              "type": "string",
              "example": "821528"
            },
            "email": {
              "type": "string",
              "example": "821528"
            },
            "title": {
              "type": "string",
              "example": "821528"
            },
            "salutation": {
              "type": "string",
              "example": "821528"
            },
            "birthday": {
              "type": "string",
              "example": "821528"
            },
            "private_street": {
              "type": "string",
              "example": "821528"
            },
            "private_street_number": {
              "type": "string",
              "example": "821528"
            },
            "private_zip_code": {
              "type": "string",
              "example": "821528"
            },
            "private_town": {
              "type": "string",
              "example": "821528"
            },
            "private_state": {
              "type": "string",
              "example": "821528"
            },
            "private_country": {
              "type": "string",
              "example": "821528"
            },
            "phone": {
              "type": "string",
              "example": "821528"
            },
            "fax": {
              "type": "string",
              "example": "821528"
            },
            "private_phone": {
              "type": "string",
              "example": "821528"
            },
            "private_mobile_phone": {
              "type": "string",
              "example": "821528"
            },
            "private_email": {
              "type": "string",
              "example": "821528"
            }
          }
        }
      },
      "createPersonRequest": {
        "type": "object",
        "required": [
          "name",
          "firstname"
        ],
        "properties": {
          "firstname": {
            "type": "string",
            "example": "John"
          },
          "name": {
            "type": "string",
            "example": "Doe"
          },
          "email": {
            "type": "string",
            "example": "jdoe@mail.com"
          }
        }
      },
      "createOrganizationRequest": {
        "type": "object",
        "required": [
          "name",
          "email"
        ],
        "properties": {
          "name": {
            "type": "string",
            "example": "Company Ltd."
          },
          "email": {
            "type": "string",
            "example": "info@company.com"
          }
        }
      },
      "createPersonResponse": {
        "type": "object",
        "required": [
          "firstName",
          "lastName"
        ],
        "properties": {}
      },
...
    }
  }
}